# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fluxa API
  version: 1.0.0
  description: 'Fluxa is a multi-currency digital asset platform built on Stellar. It provides

    wallets, instant transfers, FX conversions, scheduled recurring payments,

    batch disbursements, fiat on/off-ramps, and a developer-friendly webhook system.


    All endpoints under `/v1` require a Bearer API key in the `Authorization` header.


    ## API key scopes


    An API key can be limited to specific resources and actions. Scopes have the

    form `<resource>:<action>`, where action is `read` or `write`:


    | Resource | Scopes | Covers |

    |---|---|---|

    | wallets | `wallets:read`, `wallets:write` | `/v1/wallets` (create, trustlines, contract wallet changes) |

    | transfers | `transfers:read`, `transfers:write` | `/v1/transfers`, `/v1/transactions`, `/v1/schedules`, `/v1/refunds` |

    | batches | `batches:read`, `batches:write` | `/v1/transfers/batch` |

    | webhooks | `webhooks:read`, `webhooks:write` | `/v1/webhooks` |

    | reports | `reports:read`, `reports:write` | `/v1/usage` (`reports:write` is reserved for report generation) |


    Other scopes: `keys:*`, `audit:read`, `fiat:*`, `fx:*`, `fees:read`,

    `beneficiaries:*`, `compliance:*`. `<resource>:*` grants both actions on one

    resource and `*` grants everything.


    On scoped resource groups, `GET`/`HEAD` need the `read` scope and every other

...

...

...

    path only; never the key itself).


...

...

...

...

...

...

    '
  contact:
...
  license:
...
servers:
...
security:
- BearerAuth: []
paths:
...
  /v1/wallets/{id}/deposit/fiat:
    post:
      operationId: depositFiat
      summary: Initiate a fiat deposit
      description: 'Starts a fiat-to-crypto deposit via the configured payment rail (e.g. Flutterwave).

        Returns a payment link the customer should be redirected to.

        '
      tags:
      - Wallets
      parameters:
      - $ref: '#/components/parameters/WalletID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FiatDepositRequest'
            example:
              amount: '100.00'
              currency: NGN
              email: customer@acme.com
              name: John Doe
      responses:
        '200':
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
              example:
                payment_link: https://checkout.flutterwave.com/pay/abc123
                reference: DEP-a1b2c3d4
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/wallets/{id}/withdraw/fiat:
    post:
      operationId: withdrawFiat
      summary: Initiate a fiat withdrawal
      description: Converts USDC from the wallet and sends fiat to a bank account.
      tags:
      - Wallets
      parameters:
      - $ref: '#/components/parameters/WalletID'
      - name: X-Idempotency-Key
        in: header
        required: false
        schema:
          type: string
        description: Optional idempotency key to prevent duplicate withdrawals on network retries
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
          format: uuid
        description: Optional idempotency key to prevent duplicate withdrawals on network retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FiatWithdrawalRequest'
            example:
              amount: '50.00'
              currency: NGN
              account_bank: GTBank
              account_number: 0123456789
      responses:
        '200':
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
              example:
                reference: WIT-e5f6g7h8
                status: pending
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/withdrawals:
    post:
      operationId: initiateWithdrawal
      summary: Initiate a withdrawal
      description: Converts USDC from the wallet and sends fiat to a bank account, protected by idempotency.
      tags:
      - Wallets
      parameters:
      - name: X-Idempotency-Key
        in: header
        required: false
        schema:
          type: string
        description: Optional idempotency key to prevent duplicate withdrawals on network retries
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
          format: uuid
        description: Optional idempotency key to prevent duplicate withdrawals on network retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - wallet_id
              - amount
              - currency
              - account_bank
              - account_number
              properties:
                wallet_id:
                  type: string
                  format: uuid
                amount:
                  type: string
                currency:
                  type: string
                account_bank:
                  type: string
                account_number:
                  type: string
      responses:
        '200':
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/fiat/quote:
    post:
      operationId: previewFiat
      summary: Preview a fiat deposit or withdrawal
      description: Get a read-only quote before creating a fiat deposit or withdrawal.
      tags:
      - Wallets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - side
              - amount
              - currency
              - country
              properties:
                side:
                  type: string
                  enum:
                  - deposit
                  - withdraw
                amount:
                  type: string
                currency:
                  type: string
                country:
                  type: string
      responses:
        '200':
          description: Preview quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiatQuotePreviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/webhooks/fiat/{provider}:
    post:
      operationId: handleFiatWebhook
      summary: Handle a fiat provider webhook
      description: 'Receives webhook callbacks from fiat payment providers (e.g. Flutterwave).

        This is an internal endpoint called by the provider, not by the tenant.

        '
      tags:
      - System
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
        description: Provider name (e.g. flutterwave)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Provider-specific webhook payload
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: 'Pass a tenant API key (`sk_live_...` or `sk_test_...`) or JWT bearer token.

        In Swagger UI, enter your token in the Authorize dialog (the ''Bearer '' prefix is added automatically).

        '
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Pass an API key as `Bearer <key>` in the `Authorization` header.

        '
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: admin token
      description: Administrative bearer token.
  parameters:
    WalletID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Wallet ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: invalid request body
              status: 400
              request_id: req-example
    Unauthorized:
      description: Unauthorized — missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: missing or invalid authorization header
              status: 401
              request_id: req-example
...
    Conflict:
      description: Conflict — a request with this Idempotency-Key is still in flight
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: REQUEST_IN_PROGRESS
              message: a request with this Idempotency-Key is still in progress
              status: 409
              request_id: req-example
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: an unexpected error occurred
              status: 500
              request_id: req-example
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          - request_id
          properties:
            code:
              type: string
              enum:
              - BAD_REQUEST
              - ANCHOR_DEPOSIT_FAILED
              - ANCHOR_REGISTRATION_FAILED
              - ANCHOR_TRANSACTION_NOT_FOUND
              - ANCHOR_WITHDRAWAL_FAILED
              - API_KEY_NOT_FOUND
              - API_KEY_REVOKED
              - API_KEY_EXPIRED
              - BENEFICIARY_EXISTS
              - BENEFICIARY_NOT_ACTIVE
              - BENEFICIARY_NOT_ALLOWED
              - CONFIG_NOT_FOUND
              - CONFLICT
              - COOLING_OFF
              - CROSS_MODE_KEY_OPERATION
              - ENVIRONMENT_REQUIRED
              - IDEMPOTENCY_KEY_REQUIRED
              - IDEMPOTENCY_RECORD_PERSISTENCE_FAILED
              - IDEMPOTENCY_RESPONSE_TOO_LARGE
              - INSUFFICIENT_SCOPE
              - INVALID_ENVIRONMENT_MODE
              - INVALID_PREDICATE
              - INVALID_REQUEST
              - LAST_ORG_OWNER
              - METHOD_NOT_ALLOWED
              - MISSING_TRUSTLINE
              - NOT_IMPLEMENTED
              - PAYMENT_LINK_UNAVAILABLE
              - PREDICATE_NOT_SATISFIABLE
              - RATE_LIMITED
              - REFUND_CONFLICT
              - TEST_MODE_REQUIRED
              - TRANSFER_ID_INVALID
              - TRANSFER_ID_REQUIRED
              - TRANSFER_NOT_CANCELLABLE
              - UNSUPPORTED_FIAT_CURRENCY
              - WALLET_ID_REQUIRED
              - NOT_FOUND
              - UNAUTHORIZED
              - FORBIDDEN
              - INSUFFICIENT_SCOPE
              - INTERNAL_ERROR
              - INSUFFICIENT_SWEEPABLE_BALANCE
              - QUOTE_EXPIRED
              - QUOTE_ALREADY_USED
              - REVIEW_ALREADY_DECIDED
              - CLAIMABLE_BALANCE_NOT_PENDING
              - REQUEST_IN_PROGRESS
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            status:
              type: integer
              description: HTTP status code; matches the response status
              minimum: 400
              maximum: 599
            request_id:
              type: string
              description: Effective X-Request-ID response header value
...
    FiatQuotePreviewResponse:
      type: object
      properties:
        provider:
          type: string
        fiat_amount:
          type: string
        fiat_currency:
          type: string
        usdc_amount:
          type: string
        rate:
          type: string
        fee:
          type: string
        min_limit:
          type: string
        max_limit:
          type: string
        expires_at:
          type: string
          format: date-time
    FiatDepositRequest:
      type: object
      required:
      - amount
      - currency
      - email
      - name
      properties:
        amount:
          type: string
          description: Fiat amount as decimal string
        currency:
          type: string
          description: Fiat currency code (e.g. NGN, USD, KES)
        email:
          type: string
          format: email
        name:
          type: string
    DepositResponse:
      type: object
      properties:
        payment_link:
          type: string
          format: uri
          description: URL the customer should visit to complete payment
        reference:
          type: string
          description: Unique deposit reference (DEP-XXXXXXXX)
    FiatWithdrawalRequest:
      type: object
      required:
      - amount
      - currency
      - account_bank
      - account_number
      properties:
        amount:
          type: string
          description: Fiat amount as decimal string
        currency:
          type: string
        account_bank:
          type: string
        account_number:
          type: string
    WithdrawalResponse:
      type: object
      properties:
        reference:
          type: string
          description: Unique withdrawal reference (WIT-XXXXXXXX)
        status:
          type: string
          enum:
          - pending
          - completed
          - failed
...